Extract Mersenne Twister to separated file
