+++ mersenne_twister.h
@@ -1,0 +1,8 @@
+
+#ifndef _MERSENNE_TWISTER_H
+#define _MERSENNE_TWISTER_H
+
+extern void init_genrand(unsigned long s);
+extern unsigned long genrand_int32(void);
+
+#endif
